requirements.txt
cffi==1.14.2
Jinja2==2.11.2
llvmlite==0.33.0
MarkupSafe==1.1.1
numba==0.50.1
numpy==1.19.2
Pillow==7.0.0
pycparser==2.20
resampy==0.2.2
scipy==1.5.2
six==1.15.0
SoundFile==0.10.3.post1
opencv-python
